реализованы абстрактные классы, миксины для логирования, магические методы __repr__ создан класс "Заказ".
