+++ modules/category.py
@@ -1,6 +1,6 @@
-from .utils import LogMixin
+from .utils import LogMixin, PrintMixin
 
-class Category(LogMixin):
+class Category(LogMixin, PrintMixin):
     total_categories = 0
     total_products = 0
 
@@ -18,15 +18,3 @@ class Category(LogMixin):
 
     def __str__(self):
         return f"{self.name}, {self.description}, Лог: {', '.join(self.log_messages)}"
-
-class Order(LogMixin):
-    def __init__(self):
-        super().__init__()
-        self.products_info = []
-
-    def add_product(self, product):
-        self.log(f"Добавление продукта {product.name} в заказ.")
-        self.products_info.append(product)
-
-    def __str__(self):
-        return f"Заказ, Лог: {', '.join(self.log_messages)}"
