+++ modules/order.py
@@ -1,11 +1,13 @@
-from .product import AbstractProduct
+from modules.utils import LogMixin, PrintMixin
 
-class Order(AbstractProduct):
+class Order(LogMixin, PrintMixin):
     def __init__(self):
-        super().__init__("Order", "Заказ", 0, 0)
-        self.products = []
+        super().__init__()
+        self.products_info = []
 
     def add_product(self, product):
-        self.products.append(product)
-        print(f"Добавление продукта {product.name} в заказ.")
+        self.log(f"Добавление продукта {product.name} в заказ.")
+        self.products_info.append(product)
 
+    def __str__(self):
+        return f"Заказ, Лог: {', '.join(self.log_messages)}"
